Ready for testing, getting unsupported_grant_type error when clicking the login button.
